large reorganisation of static analysis results and processing
