+++ internal/staticanalysis/parsing/babel-parser.js
