+++ internal/staticanalysis/parsing/init_js_parser.go
@@ -1,4 +1,4 @@
-package js
+package parsing
 
 import (
 	_ "embed"
