+++ internal/staticanalysis/parsing/parsing_types.go
@@ -6,15 +6,27 @@ import (
 	"github.com/ossf/package-analysis/internal/staticanalysis/token"
 )
 
+// Language represents a programming language used for parsing
+type Language string
+
+// SymbolType denotes a type of information collected during parsing. It may be a source code token (see token package), or
 type SymbolType string
 
 const (
+	JavaScript Language = "JavaScript"
+
 	Identifier SymbolType = "Identifier" // source code identifier (variable, class, function name)
 	Literal    SymbolType = "Literal"    // source code data (string, integer, floating point literals)
 	Comment    SymbolType = "Comment"    // source code comments
 	Info       SymbolType = "Info"       // information about the parsing (e.g. number of bytes read by parser)
 	Error      SymbolType = "Error"      // any error encountered by parser; some are recoverable and some are not
 )
+
+var allLanguages = []Language{JavaScript}
+
+func SupportedLanguages() []Language {
+	return allLanguages[:]
+}
 
 type ParsedIdentifier struct {
 	Type token.IdentifierType
@@ -49,11 +61,12 @@ type ParsedComment struct {
 	Pos  token.Position
 }
 
-type ParseResult struct {
+// internalResult holds intermediate data between parsing functions
+type internalResult struct {
 	ValidInput  bool
 	Identifiers []ParsedIdentifier
 	Literals    []ParsedLiteral[any]
 	Comments    []ParsedComment
 }
 
-var InvalidInput = ParseResult{ValidInput: false}
+var InvalidInput = internalResult{ValidInput: false}
