+++ internal/worker/runstatic.go
@@ -3,15 +3,12 @@ package worker
 import (
 	"fmt"
 	"os"
-	"strings"
 	"time"
 
 	"github.com/ossf/package-analysis/internal/analysis"
 	"github.com/ossf/package-analysis/internal/log"
 	"github.com/ossf/package-analysis/internal/pkgecosystem"
 	"github.com/ossf/package-analysis/internal/sandbox"
-	"github.com/ossf/package-analysis/internal/staticanalysis"
-	"github.com/ossf/package-analysis/internal/utils"
 	"github.com/ossf/package-analysis/pkg/result"
 )
 
@@ -19,29 +16,18 @@ const staticAnalysisImage = "gcr.io/ossf-malware-analysis/static-analysis"
 const staticAnalyzeBinary = "/usr/local/bin/staticanalyze"
 const resultsJSONFile = "/results.json"
 
-/*
-RunStaticAnalyses performs sandboxed static analysis on the given package.
-Use sbOpts to customise sandbox behaviour.
-If len(tasks) > 0, only the specified static analysis tasks will be performed.
-Otherwise, all available tasks [staticanalysis.AllTasks()] will be performed.
-*/
-func RunStaticAnalyses(pkg *pkgecosystem.Pkg, sbOpts []sandbox.Option, tasks ...staticanalysis.Task) (result.StaticAnalysisResults, analysis.Status, error) {
-	if len(tasks) == 0 {
-		tasks = staticanalysis.AllTasks()
-	}
-
-	log.Info("Running static analysis tasks", "tasks", tasks)
+// RunStaticAnalyses performs sandboxed static analysis on the given package.
+// Use sbOpts to customise sandbox behaviour.
+func RunStaticAnalyses(pkg *pkgecosystem.Pkg, sbOpts []sandbox.Option) (result.StaticAnalysisResults, analysis.Status, error) {
+	log.Info("Running static analysis")
 
 	startTime := time.Now()
-
-	analyses := utils.Transform(tasks, func(t staticanalysis.Task) string { return string(t) })
 
 	args := []string{
 		staticAnalyzeBinary,
 		"-ecosystem", pkg.EcosystemName(),
 		"-package", pkg.Name(),
 		"-version", pkg.Version(),
-		"-analyses", strings.Join(analyses, ","),
 		"-output", resultsJSONFile,
 	}
